pyproject.toml
[tool.poetry]
name = "entity_typing"
version = "0.1.0"
description = ""
authors = ["fracivilization <fractal.civilization@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.9.1"
flake8 = "^3.9.2"
black = "^21.7b0"
pytest = "^6.2.4"
datasets = "^1.11.0"
torch = "^1.9.0"
gin-config = "^0.4.0"
seqeval = "^1.2.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
